(Deposit/Withdraw) Remove Legacy Bridge Flow (#3758)
* feat: remove old bridge flow

* feat: simplify provider

* fix: remove unused var

* fix: build

* fix: build

* fix: remove unused translations
