+++ packages/bridge/src/skip/index.ts
@@ -549,7 +549,7 @@ export class SkipBridgeProvider implements BridgeProvider {
   }
 
   private getViemProvider(chainID: string) {
-    const evmChain = Object.values(EthereumChainInfo).find(
+    const evmChain = EthereumChainInfo.find(
       (chain) => chain.id.toString() === chainID
     );
 
@@ -748,7 +748,7 @@ export class SkipBridgeProvider implements BridgeProvider {
     txData: BridgeTransactionRequest & { fallbackGasLimit?: number }
   ) {
     if (txData.type === "evm") {
-      const evmChain = Object.values(EthereumChainInfo).find(
+      const evmChain = EthereumChainInfo.find(
         ({ id: chainId }) => chainId === params.fromChain.chainId
       );
