+++ packages/utils/src/ethereum.ts
@@ -27,196 +27,132 @@ import {
 export const NativeEVMTokenConstantAddress =
   "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE";
 
-/** Human-displayable global source chain identifiers.
- *  TODO: use global chain IDs instead of display names as keys
- *
- * @deprecated
- */
-export type AxelarSourceChain =
-  | "Bitcoin"
-  | "Bitcoin Testnet"
-  | "Aurora Testnet"
-  | "Avalanche"
-  | "Avalanche Fuji Testnet"
-  | "Binance Smart Chain"
-  | "BSC Testnet"
-  | "Ethereum"
-  | "Goerli Testnet"
-  | "Fantom"
-  | "Fantom Testnet"
-  | "Moonbeam"
-  | "Moonbase Alpha"
-  | "Polygon"
-  | "Mumbai"
-  | "Filecoin"
-  | "Filecoin Hyperspace"
-  | "Arbitrum"
-  | "Base"
-  | "Blast"
-  | "Optimism";
-
-// TODO: maybe we can use EVM chain ID (numeric) or ethereum chain registry
-const createEthereumChainInfo = <
-  Dict extends Partial<
-    Record<
-      AxelarSourceChain,
-      Chain & {
-        chainName: AxelarSourceChain;
-        clientChainId: string;
-      }
-    >
-  >
->(
-  dict: Dict
-) => dict;
-
 function mapChainInfo<Chain>({
   chain,
-  axelarChainName: chainName,
   clientChainId,
   color,
   relativeLogoUrl,
 }: {
   chain: Chain;
-  axelarChainName: AxelarSourceChain;
   clientChainId: string;
   color: string;
   relativeLogoUrl: string;
 }) {
   return {
     ...chain,
-    chainName: chainName,
     clientChainId: clientChainId,
     color,
     relativeLogoUrl,
   };
 }
 
-export const EthereumChainInfo = createEthereumChainInfo({
-  Ethereum: mapChainInfo({
+export const EthereumChainInfo = [
+  mapChainInfo({
     chain: mainnet,
-    axelarChainName: "Ethereum",
     clientChainId: "Ethereum Main Network",
     color: "#454973",
     relativeLogoUrl: "/networks/ethereum.svg",
   }),
-  "Goerli Testnet": mapChainInfo({
+  mapChainInfo({
     chain: goerli,
-    axelarChainName: "Goerli Testnet",
     clientChainId: "Goerli Test Network",
     color: "#454973",
     relativeLogoUrl: "/networks/ethereum.svg",
   }),
-  "Binance Smart Chain": mapChainInfo({
+  mapChainInfo({
     chain: bsc,
-    axelarChainName: "Binance Smart Chain",
     clientChainId: "Binance Smart Chain Mainnet",
     color: "#f3b90c",
     relativeLogoUrl: "/networks/binance.svg",
   }),
-  "BSC Testnet": mapChainInfo({
+  mapChainInfo({
     chain: bscTestnet,
-    axelarChainName: "BSC Testnet",
     clientChainId: "Binance Smart Chain Testnet",
     color: "#f3b90c",
     relativeLogoUrl: "/networks/binance.svg",
   }),
-  Polygon: mapChainInfo({
+  mapChainInfo({
     chain: polygon,
-    axelarChainName: "Polygon",
     clientChainId: "Polygon Mainnet",
     color: "#8247E5",
     relativeLogoUrl: "/networks/polygon.svg",
   }),
-  Mumbai: mapChainInfo({
+  mapChainInfo({
     chain: polygonMumbai,
-    axelarChainName: "Mumbai",
     clientChainId: "Mumbai",
     color: "#8247E5",
     relativeLogoUrl: "/networks/polygon.svg",
   }),
-  Moonbeam: mapChainInfo({
+  mapChainInfo({
     chain: moonbeam,
-    axelarChainName: "Moonbeam",
     clientChainId: "Moonbeam Mainnet",
     color: "#FA047C",
     relativeLogoUrl: "/networks/moonbeam.svg",
   }),
-  "Moonbase Alpha": mapChainInfo({
+  mapChainInfo({
     chain: moonbaseAlpha,
-    axelarChainName: "Moonbase Alpha",
     clientChainId: "Moonbase Alpha",
     color: "#FA047C",
     relativeLogoUrl: "/networks/moonbeam.svg",
   }),
-  Fantom: mapChainInfo({
+  mapChainInfo({
     chain: fantom,
-    axelarChainName: "Fantom",
     clientChainId: "Fantom Opera",
     color: "#1b6cfc",
     relativeLogoUrl: "/networks/fantom.svg",
   }),
-  "Fantom Testnet": mapChainInfo({
+  mapChainInfo({
     chain: fantomTestnet,
-    axelarChainName: "Fantom Testnet",
     clientChainId: "Fantom Testnet",
     color: "#1b6cfc",
     relativeLogoUrl: "/networks/fantom.svg",
   }),
-  "Avalanche Fuji Testnet": mapChainInfo({
+  mapChainInfo({
     chain: avalancheFuji,
-    axelarChainName: "Avalanche Fuji Testnet",
     clientChainId: "Avalanche Fuji Testnet",
     color: "#E84142",
     relativeLogoUrl: "/networks/avalanche.svg",
   }),
-  Avalanche: mapChainInfo({
+  mapChainInfo({
     chain: avalanche,
-    axelarChainName: "Avalanche",
     clientChainId: "Avalanche C-Chain",
     color: "#E84142",
     relativeLogoUrl: "/networks/avalanche.svg",
   }),
-  Arbitrum: mapChainInfo({
+  mapChainInfo({
     chain: arbitrum,
-    axelarChainName: "Arbitrum",
     clientChainId: "Arbitrum One",
     color: "#3D4A6B",
     relativeLogoUrl: "/networks/arbitrum.svg",
   }),
-  Filecoin: mapChainInfo({
+  mapChainInfo({
     chain: filecoin,
-    axelarChainName: "Filecoin",
     clientChainId: "Filecoin - Mainnet",
     color: "#0494fc",
     relativeLogoUrl: "/networks/filecoin.svg",
   }),
-  "Filecoin Hyperspace": mapChainInfo({
+  mapChainInfo({
     chain: filecoinHyperspace,
-    axelarChainName: "Filecoin Hyperspace",
     clientChainId: "Filecoin Hyperspace",
     color: "#0494fc",
     relativeLogoUrl: "/networks/filecoin.svg",
   }),
-  Base: mapChainInfo({
+  mapChainInfo({
     chain: base as Chain,
-    axelarChainName: "Base",
     clientChainId: "Base",
     color: "#0052FF",
     relativeLogoUrl: "/networks/base.svg",
   }),
-  Blast: mapChainInfo({
+  mapChainInfo({
     chain: blast as Chain,
-    axelarChainName: "Blast",
     clientChainId: "Blast",
     color: "#fcfc03",
     relativeLogoUrl: "/networks/blast.svg",
   }),
-  Optimism: mapChainInfo({
+  mapChainInfo({
     chain: optimism as Chain,
-    axelarChainName: "Optimism",
     clientChainId: "Optimism",
     color: "#FF0420",
     relativeLogoUrl: "/networks/optimism.svg",
   }),
-});
+] as const;
