--- packages/web/integrations/axelar/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from "./use-axelar-deposit-address";
-export * from "./use-transfer-fee-query";
