--- packages/web/integrations/bridge.ts
@@ -1,8 +1,0 @@
-import { WalletKey } from "~/integrations/wallets";
-
-import { AxelarBridgeConfig } from "./axelar";
-
-export type OriginBridgeInfo = {
-  bridge: "axelar" | "nomic";
-  wallets: WalletKey[];
-} & AxelarBridgeConfig;
