+++ packages/web/integrations/index.ts
@@ -1,4 +1,3 @@
-export * from "./bridge";
 export * from "./fiat-ramps";
 export * from "./use-client-tx-event-toasts";
 export * from "./wallets";
