+++ packages/web/localizations/es.json
@@ -171,12 +171,6 @@
       "FeeAPY": "Tarifa TAI",
       "liquidity": "Mi Liquidez"
     },
-    "selectAssetSource": {
-      "installWallet": "Instalar {walletName}",
-      "next": "Siguiente",
-      "titleDeposit": "Depositar desde",
-      "titleWithdraw": "Retirar a"
-    },
     "stakedAssets": "OSMO Abonado",
     "table": {
       "columns": {
@@ -189,11 +183,6 @@
       "depositButton": "Depositar",
       "hideZero": "Ocultar saldos cero",
       "hideBalances": "Saldos enmascarados",
-      "preTransfer": {
-        "currentBal": "Saldo actual",
-        "deposit": "Depositar",
-        "withdraw": "Retirar"
-      },
       "search": "Buscar",
       "sort": {
         "balance": "Balance",
@@ -217,58 +206,20 @@
       "availableOn": "Disponible en {network}",
       "approving": "Aprobando...",
       "approveInWallet": "Aprobar en billetera",
-      "enterAddress": "dirección {network}",
       "errors": {
-        "seeRequest": "Ver solicitud en {walletName}",
-        "invalidAddress": "Dirección inválida"
+        "seeRequest": "Ver solicitud en {walletName}"
       },
       "from": "De",
       "loading": "Cargando",
-      "titleDeposit": "Depositar {coinDenom}",
-      "titleWithdraw": "Retirar {coinDenom}",
       "to": "A",
       "transferFee": "Cuota de transferencia",
-      "waitTime": "{minutes} minutos",
-      "warnDepositAddressDifferent": "Advertencia: la cuenta seleccionada en {address} difiere de la cuenta con la que se depositó por última vez.",
       "expectedOutput": "Rendimiento esperado",
       "expectedOutputInfo": "Dependiendo del proveedor utilizado, puede haber una producción menor según las tarifas, el deslizamiento y el impacto en el precio.",
       "priceImpact": "Impacto en el precio",
       "sending": "Enviando...",
       "transferAnyway": "Transferir de todos modos"
     },
-    "transferAssetSelect": {
-      "buttonNext": "Siguiente",
-      "deposit": "Depositar",
-      "network": "Red",
-      "withdraw": "Retirar"
-    },
     "unbondedAssets": "No abonado",
-    "nomic": {
-      "addresses": "direcciones.",
-      "bitcoin": "Bitcoin (BTC)",
-      "bitcoinMinerFee": "Tarifa de minero de Bitcoin",
-      "bridgeAtCapacity": "El puente se encuentra actualmente al límite de su capacidad. Por favor, inténtelo de nuevo más tarde.",
-      "bridgeFee": "Tarifa Puente Nómica",
-      "confirmations": "6 confirmaciones",
-      "enterBitcoinAddress": "Ingrese la dirección de Bitcoin",
-      "estimatedArrival": "llegada estimada",
-      "estimatedTimeTag": "Hora prevista",
-      "estimatedTime": "5 minutos",
-      "expirationTime": "4 dias",
-      "learnMore": "Obtenga más información en",
-      "lossOfFunds": "El uso de la dirección después de este tiempo resultará en la pérdida de fondos.",
-      "minDeposit": "Depósito mínimo",
-      "minWithdrawal": "Retiro Mínimo",
-      "nomicCapacityLimit": "Nomic tiene un límite de capacidad de",
-      "onlyUse": "Uso único",
-      "pendingDeposits": "Depósitos pendientes",
-      "poweredBy": "Energizado por",
-      "proceed": "Proceder",
-      "depositExpiration": "Esta dirección de depósito vence en",
-      "undergoAudit": "mientras se somete a auditoría (finalización estimada en enero de 2024).",
-      "warning": "ADVERTENCIA:",
-      "loading": "Cargando..."
-    },
     "activateUnverifiedAssetsModal": {
       "activate": "Activar activos no verificados",
       "description": "es un token no verificado. ¿Quieres activarlo? Asegúrese de investigar a fondo antes de operar.",
@@ -1164,7 +1115,6 @@
       "veryHigh": "Muy alto"
     }
   },
-  "disclaimer": "Descargo de responsabilidad",
   "remaining": "restantes",
   "timeUnits": {
     "second": "segundo",
