+++ packages/web/localizations/fa.json
@@ -171,12 +171,6 @@
       "FeeAPY": "کارمزذ سپرده گذاری",
       "liquidity": "نقدینگی من"
     },
-    "selectAssetSource": {
-      "installWallet": "نصب کیف پول {walletName}",
-      "next": "بعدی",
-      "titleDeposit": "برداشت از",
-      "titleWithdraw": "پرداخت به"
-    },
     "stakedAssets": "سپرده گذاری شده OSMO",
     "table": {
       "columns": {
@@ -189,11 +183,6 @@
       "depositButton": "واریز",
       "hideZero": "مخفی کردن موجودی های صفر",
       "hideBalances": "تعادل های پوشیده شده",
-      "preTransfer": {
-        "currentBal": "موجودی فعلی",
-        "deposit": "واریز وجه",
-        "withdraw": "برداشت وجه"
-      },
       "search": "جستجو",
       "sort": {
         "balance": "موجودی",
@@ -217,58 +206,20 @@
       "availableOn": "دارایی در شبکه {network}",
       "approving": "تایید...",
       "approveInWallet": "تایید در کیف پول",
-      "enterAddress": "آدرس {network}",
       "errors": {
-        "seeRequest": "درخواست را در {walletName} ببینید",
-        "invalidAddress": "آدرس نامعتبر"
+        "seeRequest": "درخواست را در {walletName} ببینید"
       },
       "from": "از",
       "loading": "در حال بارگذاری",
-      "titleDeposit": "واریز {coinDenom}",
-      "titleWithdraw": "برداشت {coinDenom}",
       "to": "به",
       "transferFee": "کارمزد انتقال",
-      "waitTime": "{minutes} دقیقه",
-      "warnDepositAddressDifferent": "هشدار: {address}  با آخرین آدرس برداشت و واریز شما متفاوت است.",
       "expectedOutput": "خروجی مورد انتظار",
       "expectedOutputInfo": "بسته به ارائه دهنده مورد استفاده، ممکن است با توجه به هزینه ها، لغزش و تاثیر قیمت، خروجی کمتری وجود داشته باشد.",
       "priceImpact": "تاثیر قیمت",
       "sending": "در حال ارسال...",
       "transferAnyway": "انتقال به هر حال"
     },
-    "transferAssetSelect": {
-      "buttonNext": "بعدی",
-      "deposit": "واریز",
-      "network": "شبکه",
-      "withdraw": "برداشت"
-    },
     "unbondedAssets": "بازکردن",
-    "nomic": {
-      "addresses": "آدرس ها.",
-      "bitcoin": "بیت کوین (BTC)",
-      "bitcoinMinerFee": "کارمزد استخراج بیت کوین",
-      "bridgeAtCapacity": "ظرفیت پل در حال حاضر به اتمام رسیده است. لطفاً بعداً دوباره امتحان کنید.",
-      "bridgeFee": "هزینه پل اسمی",
-      "confirmations": "6 تائید",
-      "enterBitcoinAddress": "آدرس بیت کوین را وارد کنید",
-      "estimatedArrival": "زمان تقریبی رسیدن به مقصد",
-      "estimatedTimeTag": "زمان تخمین زده شده",
-      "estimatedTime": "5 دقیقه",
-      "expirationTime": "4 روز",
-      "learnMore": "بیشتر بدانید در",
-      "lossOfFunds": "استفاده از آدرس پس از این زمان منجر به از دست دادن وجه می شود.",
-      "minDeposit": "حداقل سپرده",
-      "minWithdrawal": "حداقل برداشت",
-      "nomicCapacityLimit": "Nomic دارای محدودیت ظرفیت است",
-      "onlyUse": "فقط استفاده کنید",
-      "pendingDeposits": "سپرده های معلق",
-      "poweredBy": "پشتیبانی شده توسط",
-      "proceed": "ادامه دهید",
-      "depositExpiration": "این آدرس سپرده منقضی می شود",
-      "undergoAudit": "در حین انجام حسابرسی (تکمیل تخمینی ژانویه 2024).",
-      "warning": "هشدار:",
-      "loading": "در حال بارگذاری..."
-    },
     "activateUnverifiedAssetsModal": {
       "activate": "دارایی های تایید نشده را فعال کنید",
       "description": "یک توکن تایید نشده است. آیا می خواهید آن را فعال کنید؟ حتما قبل از معامله به طور کامل تحقیق کنید.",
@@ -1164,7 +1115,6 @@
       "veryHigh": "خیلی بالا"
     }
   },
-  "disclaimer": "سلب مسئولیت",
   "remaining": "باقی مانده است",
   "timeUnits": {
     "second": "دومین",
