+++ packages/web/localizations/zh-cn.json
@@ -171,12 +171,6 @@
       "FeeAPY": "费用 APY",
       "liquidity": "我的流动性"
     },
-    "selectAssetSource": {
-      "installWallet": "安装 {walletName}",
-      "next": "下一步",
-      "titleDeposit": "存入于",
-      "titleWithdraw": "提币至"
-    },
     "stakedAssets": "质押 OSMO",
     "table": {
       "columns": {
@@ -189,11 +183,6 @@
       "depositButton": "存入",
       "hideZero": "隐藏0余额",
       "hideBalances": "隐藏余额",
-      "preTransfer": {
-        "currentBal": "当前余额",
-        "deposit": "存入",
-        "withdraw": "提币"
-      },
       "search": "搜索",
       "sort": {
         "balance": "余额",
@@ -217,58 +206,20 @@
       "availableOn": "在 {network} 网络上可用",
       "approving": "正在批准...",
       "approveInWallet": "在钱包中批准",
-      "enterAddress": "{network}地址",
       "errors": {
-        "seeRequest": "在{walletName}中查看请求",
-        "invalidAddress": "无效地址"
+        "seeRequest": "在{walletName}中查看请求"
       },
       "from": "从",
       "loading": "加载中",
-      "titleDeposit": "存入 {coinDenom}",
-      "titleWithdraw": "提币 {coinDenom}",
       "to": "至",
       "transferFee": "转账费",
-      "waitTime": "{minutes} 分钟",
-      "warnDepositAddressDifferent": "警告：{address}中选定的帐户与您上次存入的帐户不同.",
       "expectedOutput": "预期输出",
       "expectedOutputInfo": "根据所使用的提供商的不同，根据费用、滑点和价格影响，产出可能会较低。",
       "priceImpact": "价格影响",
       "sending": "正在发送...",
       "transferAnyway": "无论如何转移"
     },
-    "transferAssetSelect": {
-      "buttonNext": "下一步",
-      "deposit": "存入",
-      "network": "网络",
-      "withdraw": "提币"
-    },
     "unbondedAssets": "未绑定资产",
-    "nomic": {
-      "addresses": "地址。",
-      "bitcoin": "比特币（BTC）",
-      "bitcoinMinerFee": "比特币矿工费",
-      "bridgeAtCapacity": "该桥目前已满负荷。请稍后再试。",
-      "bridgeFee": "诺米克桥费",
-      "confirmations": "6 确认",
-      "enterBitcoinAddress": "输入比特币地址",
-      "estimatedArrival": "预计抵达时间",
-      "estimatedTimeTag": "预计时间",
-      "estimatedTime": "5分钟",
-      "expirationTime": "4天",
-      "learnMore": "了解更多信息，请访问",
-      "lossOfFunds": "在此时间之后使用地址将导致资金损失。",
-      "minDeposit": "最低存款额",
-      "minWithdrawal": "最低提款额",
-      "nomicCapacityLimit": "Nomic 的容量限制为",
-      "onlyUse": "仅使用",
-      "pendingDeposits": "待处理存款",
-      "poweredBy": "供电",
-      "proceed": "继续",
-      "depositExpiration": "该充值地址到期时间为",
-      "undergoAudit": "正在接受审核（预计于 2024 年 1 月完成）。",
-      "warning": "警告：",
-      "loading": "加载中..."
-    },
     "activateUnverifiedAssetsModal": {
       "activate": "激活未验证资产",
       "description": "是一个未经验证的令牌。您想激活它吗？交易前一定要彻底研究。",
@@ -1164,7 +1115,6 @@
       "veryHigh": "很高"
     }
   },
-  "disclaimer": "免责声明",
   "remaining": "其余的",
   "timeUnits": {
     "second": "第二",
