+++ packages/web/modals/index.ts
@@ -1,8 +1,6 @@
 export * from "./activate-unverified-token-confirmation";
 export * from "./add-liquidity";
 export * from "./base";
-export * from "./bridge-transfer-v1";
-export * from "./bridge-transfer-v2";
 export * from "./create-pool";
 export * from "./external-links-modal";
 export * from "./fiat-on-ramp-selection";
@@ -11,13 +9,10 @@ export * from "./ibc-transfer";
 export * from "./lock-tokens";
 export * from "./menu-options";
 export * from "./one-click-trading-intro-modal";
-export * from "./pre-transfer";
 export * from "./remove-liquidity";
-export * from "./select-asset-source";
 export * from "./settings";
 export * from "./settings";
 export * from "./superfluid-validator";
 export * from "./token-select";
 export * from "./trade-tokens";
-export * from "./transfer-asset-select";
 export * from "./wallet-select";
