+++ packages/web/stores/assets/index.ts
@@ -1,3 +1,2 @@
 export * from "./assets-store";
-export * from "./transfer-ui-config";
 export * from "./types";
