+++ packages/web/stores/assets/types.ts
@@ -2,7 +2,6 @@ import { AppCurrency } from "@keplr-wallet/types";
 import { CoinPretty, PricePretty } from "@keplr-wallet/unit";
 
 import { FiatRampKey } from "~/integrations";
-import { OriginBridgeInfo } from "~/integrations";
 
 export interface IBCAsset {
   counterpartyChainId: string;
@@ -23,9 +22,6 @@ export interface IBCAsset {
   // If this is a multihop ibc, need to special case because the denom on osmosis
   // isn't H(source_denom), but rather H(ibc_path)
   ibcTransferPathDenom?: string;
-
-  /** Additional info to support non-IBC bridge integration. */
-  originBridgeInfo?: OriginBridgeInfo;
 
   /** Keys for fiat on/off ramps. Ramp must accept asset's major denom (e.g. `ATOM`). */
   fiatRamps?: { rampKey: FiatRampKey; assetKey: string }[];
@@ -65,7 +61,6 @@ export interface IBCBalance extends CoinBalance {
   destChannelId: string;
   isUnstable?: boolean;
   isVerified: boolean;
-  originBridgeInfo?: OriginBridgeInfo;
   fiatRamps?: { rampKey: FiatRampKey; assetKey: string }[];
 }
